# backend/src/main/resources/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
...
paths:
...
  /events/{id}/location/{propertyId}:
    get:
      tags:
        - event-property-reference-controller
      description: get-location-by-event-Id
      operationId: followPropertyReference-event-get
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: propertyId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelLocation'
        '404':
          description: Not Found
    delete:
      tags:
        - event-property-reference-controller
      description: delete-location-by-event-Id
      operationId: deletePropertyReferenceId-event-delete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: propertyId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: No Content
        '404':
          description: Not Found
  /locations:
    get:
      tags:
        - location-entity-controller
      description: get-location
      operationId: getCollectionResource-location-get_1
      responses:
        '200':
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelEntityModelLocation'
            application/x-spring-data-compact+json:
              schema:
                $ref: '#/components/schemas/CollectionModelEntityModelLocation'
            text/uri-list:
              schema:
                type: string
    post:
      tags:
        - location-entity-controller
      description: create-location
      operationId: postCollectionResource-location-post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LocationRequestBody'
        required: true
      responses:
        '201':
          description: Created
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelLocation'
  /locations/{id}:
    get:
      tags:
        - location-entity-controller
      description: get-location
      operationId: getItemResource-location-get
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelLocation'
        '404':
          description: Not Found
    put:
      tags:
        - location-entity-controller
      description: update-location
      operationId: putItemResource-location-put
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LocationRequestBody'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelLocation'
        '201':
          description: Created
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelLocation'
        '204':
          description: No Content
    delete:
      tags:
        - location-entity-controller
      description: delete-location
      operationId: deleteItemResource-location-delete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: No Content
        '404':
          description: Not Found
    patch:
      tags:
        - location-entity-controller
      description: patch-location
      operationId: patchItemResource-location-patch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LocationRequestBody'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelLocation'
        '204':
          description: No Content
...
  /rest/locations:
    get:
      tags:
        - Location
      summary: Получить список всех локаций
      operationId: getAll
      responses:
        '200':
          description: Успешно
          content:
            application/hal+json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Location'
        '500':
          description: Internal Server Error
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/Exception'
  /rest/locations/{id}:
    get:
      tags:
        - Location
      summary: Получить локацию
      operationId: getById_1
      parameters:
        - name: id
          in: path
          description: Идентификатор локации
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Успешно
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/Location'
        '404':
          description: Локация не найдена
        '500':
          description: Internal Server Error
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/Exception'
  /rest/locations/{id}/events:
    get:
      tags:
        - Location
      summary: Получить список всех событий локации
      operationId: getEvents
      parameters:
        - name: id
          in: path
          description: Идентификатор локации
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Успешно
          content:
            application/hal+json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
        '500':
          description: Internal Server Error
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/Exception'
...
components:
  schemas:
    Exception:
      type: object
      properties:
        message:
          type: string
...
    Links:
      type: object
      additionalProperties:
        $ref: '#/components/schemas/Link'
...
    Event:
      type: object
      properties:
        id:
          type: integer
          format: int64
        creationDate:
          type: string
          format: date-time
        startDate:
          type: string
          format: date-time
        duration:
          type: integer
          format: int64
        totalMembers:
          type: integer
          format: int32
        countMembers:
          type: integer
          format: int32
        type:
          $ref: '#/components/schemas/EventType'
        location:
          $ref: '#/components/schemas/Location'
    EventType:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
    Location:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        latitude:
          type: number
          format: float
        longitude:
          type: number
          format: float
        address:
          type: string
...
    EntityModelLocation:
      type: object
      properties:
        title:
          type: string
        latitude:
          type: number
          format: float
        longitude:
          type: number
          format: float
        address:
          type: string
        _links:
          $ref: '#/components/schemas/Links'
...
    CollectionModelEntityModelLocation:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            locations:
              type: array
              items:
                $ref: '#/components/schemas/EntityModelLocation'
        _links:
          $ref: '#/components/schemas/Links'
...
    LocationRequestBody:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        latitude:
          type: number
          format: float
        longitude:
          type: number
          format: float
        address:
          type: string
...
    Link:
      type: object
      properties:
        href:
          type: string
        hreflang:
          type: string
        title:
          type: string
        type:
          type: string
        deprecation:
          type: string
        profile:
          type: string
        name:
          type: string
        templated:
          type: boolean